lowered test requirements
