+++ tests/analog/test_ensemble.py
@@ -5,7 +5,10 @@
 #
 # Licensed under the MIT License
 
-"""Tests for deterministic unitary ensemble evolution in analog simulations."""
+"""Tests for deterministic unitary ensemble evolution in analog simulations.
+
+Includes spin-chain two-time correlator checks against dense ED (see ``mqt.yaqs.analog.ensemble``).
+"""
 
 from __future__ import annotations
 
@@ -15,11 +18,12 @@ import pytest
 from mqt.yaqs import simulator
 from mqt.yaqs.analog.ensemble import ensemble_member_worker
 from mqt.yaqs.core.data_structures.hamiltonian import Hamiltonian
+from mqt.yaqs.core.data_structures.mpo import MPO
 from mqt.yaqs.core.data_structures.mps import MPS
 from mqt.yaqs.core.data_structures.noise_model import NoiseModel
 from mqt.yaqs.core.data_structures.simulation_parameters import AnalogSimParams, EvolutionMode, Observable
 from mqt.yaqs.core.data_structures.state import State
-from mqt.yaqs.core.libraries.gate_library import Z
+from mqt.yaqs.core.libraries.gate_library import BaseGate, X, Y, Z
 
 
 def test_unitary_ensemble_observable_average() -> None:
@@ -309,3 +313,187 @@ def test_list_initial_states_with_noise_raises() -> None:
         ),
     ):
         simulator.run(initial_states, hamiltonian, sim_params, noise_model=noise_model, parallel=False)
+
+
+# --- Spin-chain ED helpers and integration tests ---
+
+
+def _embed_one_site_operator(length: int, site: int, op2: np.ndarray) -> np.ndarray:
+    left_dim = 2**site
+    right_dim = 2 ** (length - site - 1)
+    return np.kron(np.kron(np.eye(left_dim, dtype=np.complex128), op2), np.eye(right_dim, dtype=np.complex128))
+
+
+def _embed_two_site_operator(length: int, site_left: int, op4: np.ndarray) -> np.ndarray:
+    left_dim = 2**site_left
+    right_dim = 2 ** (length - site_left - 2)
+    return np.kron(np.kron(np.eye(left_dim, dtype=np.complex128), op4), np.eye(right_dim, dtype=np.complex128))
+
+
+def _spin_current_bond_matrix(j_coupling: float) -> np.ndarray:
+    x = np.array([[0.0, 1.0], [1.0, 0.0]], dtype=np.complex128)
+    y = np.array([[0.0, -1.0j], [1.0j, 0.0]], dtype=np.complex128)
+    return 0.25 * j_coupling * (np.kron(x, y) - np.kron(y, x))
+
+
+def _periodic_bond_endpoints(length: int) -> list[tuple[int, int]]:
+    return [(i, (i + 1) % length) for i in range(length)]
+
+
+def _spin_current_observable_for_periodic_bond(site_a: int, site_b: int, j_xy: float) -> Observable:
+    return Observable(BaseGate(_spin_current_bond_matrix(j_xy)), sites=[site_a, site_b])
+
+
+def _build_xxz_tf_open_chain_dense(length: int, j_xy: float, delta: float, h_x: float) -> np.ndarray:
+    x = np.array([[0.0, 1.0], [1.0, 0.0]], dtype=np.complex128)
+    y = np.array([[0.0, -1.0j], [1.0j, 0.0]], dtype=np.complex128)
+    z = np.array([[1.0, 0.0], [0.0, -1.0]], dtype=np.complex128)
+    xx = np.kron(x, x)
+    yy = np.kron(y, y)
+    zz = np.kron(z, z)
+    dim = 2**length
+    h_dense = np.zeros((dim, dim), dtype=np.complex128)
+    for i in range(length - 1):
+        h_dense += 0.25 * j_xy * _embed_two_site_operator(length, i, xx)
+        h_dense += 0.25 * j_xy * _embed_two_site_operator(length, i, yy)
+        h_dense += 0.25 * delta * _embed_two_site_operator(length, i, zz)
+    for r in range(length):
+        h_dense += 0.5 * h_x * _embed_one_site_operator(length, r, x)
+    return h_dense
+
+
+def _ed_first_k_basis_two_time_means(
+    *,
+    length: int,
+    j_xy: float,
+    delta: float,
+    h_x: float,
+    times: np.ndarray,
+    k: int,
+    site: int,
+    probes: list[tuple[np.ndarray, np.ndarray]],
+) -> list[np.ndarray]:
+    dim = 2**length
+    k_eff = min(k, dim)
+    h_dense = _build_xxz_tf_open_chain_dense(length, j_xy, delta, h_x)
+    evals, evecs = np.linalg.eigh(h_dense)
+    evecs_h = evecs.conj().T
+
+    embedded_b = [_embed_one_site_operator(length, site, b2) for _a2, b2 in probes]
+    embedded_a = [_embed_one_site_operator(length, site, a2) for a2, _b2 in probes]
+
+    out = np.zeros((len(probes), len(times)), dtype=np.complex128)
+    for t_idx, t in enumerate(times):
+        phases = np.exp(-1j * evals * t)
+        u_t = (evecs * phases[np.newaxis, :]) @ evecs_h
+        for p_idx, (a_op, b_op) in enumerate(zip(embedded_a, embedded_b, strict=True)):
+            m = u_t.conj().T @ a_op @ u_t @ b_op
+            out[p_idx, t_idx] = np.sum(np.diag(m)[:k_eff]) / float(k_eff)
+
+    return [out[p_idx] for p_idx in range(len(probes))]
+
+
+def test_xxz_transverse_unitary_ensemble_pauli_and_two_time_vs_ed() -> None:
+    """Match YAQS two-time correlators against dense ED for a transverse-field XXZ setup.
+
+    Parameters are tuned for CI runtime while keeping max error below ``1e-5`` against ED:
+    ``k=16`` ensemble members (was 64) and ``parallel=True`` (unitary, deterministic).
+    """
+    length = 6
+    j_xy = 1.0
+    delta = 0.7
+    h_x = 0.5
+    t_final = 2.0
+    dt = 0.05
+    k = 16
+    mid = length // 2
+
+    h_mpo = Hamiltonian.pauli(
+        length=length,
+        two_body=[(0.25 * j_xy, "X", "X"), (0.25 * j_xy, "Y", "Y"), (0.25 * delta, "Z", "Z")],
+        one_body=[(0.5 * h_x, "X")],
+        bc="open",
+    )
+    states = [State(length, initial="basis", basis_string=format(i, f"0{length}b")) for i in range(k)]
+
+    ox = Observable(X(), mid)
+    oy = Observable(Y(), mid)
+    oz = Observable(Z(), mid)
+    pairs = [(ox, ox), (oy, oy), (oz, oz), (oz, ox)]
+
+    sim_params = AnalogSimParams(
+        observables=[],
+        elapsed_time=t_final,
+        dt=dt,
+        max_bond_dim=256,
+        threshold=1e-12,
+        sample_timesteps=True,
+        show_progress=False,
+        multi_time_observables=pairs,
+    )
+    simulator.run(states, h_mpo, sim_params, noise_model=None, parallel=True)
+    assert sim_params.multi_time_observables_results is not None
+    assert sim_params.multi_time_observables_times is not None
+    yaqs = np.asarray(sim_params.multi_time_observables_results, dtype=np.complex128)
+    times = np.asarray(sim_params.multi_time_observables_times, dtype=np.float64)
+
+    x2 = np.array([[0.0, 1.0], [1.0, 0.0]], dtype=np.complex128)
+    y2 = np.array([[0.0, -1.0j], [1.0j, 0.0]], dtype=np.complex128)
+    z2 = np.array([[1.0, 0.0], [0.0, -1.0]], dtype=np.complex128)
+
+    ed_series = _ed_first_k_basis_two_time_means(
+        length=length,
+        j_xy=j_xy,
+        delta=delta,
+        h_x=h_x,
+        times=times,
+        k=k,
+        site=mid,
+        probes=[(x2, x2), (y2, y2), (z2, z2), (z2, x2)],
+    )
+    ed = np.vstack(ed_series)
+
+    max_abs_per_pair = np.max(np.abs(yaqs - ed), axis=1)
+    assert np.all(max_abs_per_pair < 1e-5), f"pairwise max errors were {max_abs_per_pair!r}"
+
+
+def test_two_time_correlator_probe_row_diagonal_matches_expectation_at_t0() -> None:
+    """At ``t=0``, two-time block diagonal entry ``(r,r)`` must match dense ``<psi|j_r j_r|psi>``."""
+    length = 4
+    j_xy = 1.0
+    delta = 0.8
+    mps = MPS(length, state="random", pad=8)
+    mps.normalize("B")
+
+    row = tuple(_spin_current_observable_for_periodic_bond(a, b, j_xy) for a, b in _periodic_bond_endpoints(length))
+    s_index = 1
+    obs_s = row[s_index]
+    pairs = [(obs_r, obs_s) for obs_r in row]
+
+    h = MPO.pauli(
+        length=length,
+        two_body=[(0.25 * j_xy, "X", "X"), (0.25 * j_xy, "Y", "Y"), (0.25 * delta, "Z", "Z")],
+        bc="periodic",
+    )
+
+    sim_params = AnalogSimParams(
+        observables=[],
+        elapsed_time=0.0,
+        dt=0.5,
+        max_bond_dim=32,
+        threshold=1e-10,
+        order=1,
+        sample_timesteps=True,
+        show_progress=False,
+        multi_time_observables=pairs,
+    )
+
+    _, mat = ensemble_member_worker((0, mps, sim_params, h))
+    assert mat is not None
+    val_worker = float(np.real(mat[s_index, 0]))
+
+    psi = np.asarray(mps.to_vec(), dtype=np.complex128)
+    j_bond = _spin_current_bond_matrix(j_xy)
+    jr = _embed_two_site_operator(length, 1, j_bond)
+    expected = float(np.real(np.vdot(psi, jr @ jr @ psi)))
+    assert val_worker == pytest.approx(expected, rel=0, abs=1e-6)
